+++ routes/api.php
@@ -8,7 +8,6 @@ use App\Http\Controllers\API\UserController;
 use App\Http\Controllers\API\HomeController;
 
 use App\Http\Middleware\JSONResponse;
-use App\Http\Middleware\HandleCors;
 
 
 
@@ -23,18 +22,11 @@ use App\Http\Middleware\HandleCors;
 |
 */
 
-// Route::middleware('auth:sanctum')->get('/user', function (Request $request) {
-//     return $request->user();
-// });
-
 Route::prefix('v1')->group(function () {
-
     // welcome route
     Route::get('/welcome', function () {
-        return response()->json(['message' => 'Welcome to the API!']);
+        return response()->json(['message' => 'Welcome to the DC Engine API!']);
     });
-    
-    Route::get('search', [HomeController::class, 'search']);
 
     // Authentication Routes
     Route::group([
@@ -51,13 +43,18 @@ Route::prefix('v1')->group(function () {
             Route::get('logout', [AuthController::class, 'logout']);
             Route::get('user', [AuthController::class, 'user']);
         });
-
     });
 
-    // Home Route Group
-    // Route::group(function () {
-    //     Route::get('search', [HomeController::class, 'search']);
-    // });
+    // Search Route Group
+    Route::group([
+        'prefix' => 'search',
+        // middleware
+        'middleware' => [JSONResponse::class]
+    
+    ],function () {
+        Route::get('/', [HomeController::class, 'search']);
+        Route::get('/suggestions', [HomeController::class, 'search_suggestions']);
+    });
 
     // Destinations Routes
     Route::group([
@@ -65,10 +62,9 @@ Route::prefix('v1')->group(function () {
         // middleware
         'middleware' => [JSONResponse::class]
     ], function () {
-        Route::post('/', [DestinationController::class, 'index']);
+        Route::get('/', [DestinationController::class, 'index']);
         Route::get('/{id}', [DestinationController::class, 'show']);
         Route::post('/{id}/favorite', [DestinationController::class, 'favorite'])->middleware('auth:sanctum');
-
 
         // Route::post('/', [DestinationController::class, 'store']);
         // Route::put('/{id}', [DestinationController::class, 'update']);